requirements.txt
chromadb==0.5.0
langchain==0.2.1
langchain-community==0.2.1
pypdf==4.2.0
ollama==0.2.0
tqdm==4.66.4
streamlit==1.35.0
watchdog==4.0.1
fastapi
uvicorn
sentence-transformers
einops
python-multipart
openai-whisper
torch
torchvision
torchaudio
requests